pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "speech-mine"
version = "0.1.0"
description = "A powerful tool for extracting and analyzing speech data from audio files with known speaker counts and contents."
readme = "README_PYPI.md"

license = "MIT"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.11"
dependencies = [
    "faster-whisper>=1.2.0",
    "pandas>=2.3.2",
    "pyannote-audio>=3.3.2",
    "pydub>=0.25.1",
    "pytest>=8.4.2",
    "PyYAML>=6.0.0",
    "rapidfuzz>=3.14.1",
    "tqdm>=4.67.1",
]

[project.urls]
Homepage = "https://github.com/your-org/speech-mine"
Repository = "https://github.com/your-org/speech-mine"
Issues = "https://github.com/your-org/speech-mine/issues"

[project.scripts]
speech-mine = "speech_mine.diarizer.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/speech_mine"]

[dependency-groups]
dev = [
    "mkdocs-material>=9.7.4",
    "pytest>=8.4.2",
]
